feat: parser as a struct
This makes it easier to add runtime configuration options
to the parser to enable options while building.
