+++ internal/dataplane/parser/translate_knative.go
@@ -1,0 +1,133 @@
+package parser
+
+import (
+	"fmt"
+	"sort"
+
+	"github.com/kong/go-kong/kong"
+	"github.com/kong/kubernetes-ingress-controller/v2/internal/dataplane/kongstate"
+	"github.com/kong/kubernetes-ingress-controller/v2/internal/util"
+	knative "knative.dev/networking/pkg/apis/networking/v1alpha1"
+)
+
+func (p *Parser) ingressRulesFromKnativeIngress() ingressRules {
+	result := newIngressRules()
+
+	ingressList, err := p.storer.ListKnativeIngresses()
+	if err != nil {
+		p.logger.Errorf("failed to list Knative Ingresses: %v", err)
+		return result
+	}
+
+	sort.SliceStable(ingressList, func(i, j int) bool {
+		return ingressList[i].CreationTimestamp.Before(
+			&ingressList[j].CreationTimestamp)
+	})
+
+	services := map[string]kongstate.Service{}
+	secretToSNIs := newSecretNameToSNIs()
+
+	for _, ingress := range ingressList {
+		ingressSpec := ingress.Spec
+
+		secretToSNIs.addFromIngressV1beta1TLS(knativeIngressToNetworkingTLS(ingress.Spec.TLS), ingress.Namespace)
+
+		for i, rule := range ingressSpec.Rules {
+			hosts := rule.Hosts
+			if rule.HTTP == nil {
+				continue
+			}
+			for j, rule := range rule.HTTP.Paths {
+				path := rule.Path
+
+				if path == "" {
+					path = "/"
+				}
+				r := kongstate.Route{
+					Ingress: util.FromK8sObject(ingress),
+					Route: kong.Route{
+						Name:              kong.String(fmt.Sprintf("%s.%s.%d%d", ingress.Namespace, ingress.Name, i, j)),
+						Paths:             kong.StringSlice(path),
+						StripPath:         kong.Bool(false),
+						PreserveHost:      kong.Bool(true),
+						Protocols:         kong.StringSlice("http", "https"),
+						RegexPriority:     kong.Int(0),
+						RequestBuffering:  kong.Bool(true),
+						ResponseBuffering: kong.Bool(true),
+					},
+				}
+				r.Hosts = kong.StringSlice(hosts...)
+
+				knativeBackend := knativeSelectSplit(rule.Splits)
+				serviceName := fmt.Sprintf("%s.%s.%s", knativeBackend.ServiceNamespace, knativeBackend.ServiceName,
+					knativeBackend.ServicePort.String())
+				serviceHost := fmt.Sprintf("%s.%s.%s.svc", knativeBackend.ServiceName, knativeBackend.ServiceNamespace,
+					knativeBackend.ServicePort.String())
+				service, ok := services[serviceName]
+				if !ok {
+
+					var headers []string
+					for key, value := range knativeBackend.AppendHeaders {
+						headers = append(headers, key+":"+value)
+					}
+					for key, value := range rule.AppendHeaders {
+						headers = append(headers, key+":"+value)
+					}
+
+					service = kongstate.Service{
+						Service: kong.Service{
+							Name:           kong.String(serviceName),
+							Host:           kong.String(serviceHost),
+							Port:           kong.Int(DefaultHTTPPort),
+							Protocol:       kong.String("http"),
+							Path:           kong.String("/"),
+							ConnectTimeout: kong.Int(DefaultServiceTimeout),
+							ReadTimeout:    kong.Int(DefaultServiceTimeout),
+							WriteTimeout:   kong.Int(DefaultServiceTimeout),
+							Retries:        kong.Int(DefaultRetries),
+						},
+						Namespace: ingress.Namespace,
+						Backend: kongstate.ServiceBackend{
+							Name: knativeBackend.ServiceName,
+							Port: PortDefFromIntStr(knativeBackend.ServicePort),
+						},
+					}
+					if len(headers) > 0 {
+						service.Plugins = append(service.Plugins, kong.Plugin{
+							Name: kong.String("request-transformer"),
+							Config: kong.Configuration{
+								"add": map[string]interface{}{
+									"headers": headers,
+								},
+							},
+						})
+					}
+				}
+				service.Routes = append(service.Routes, r)
+				services[serviceName] = service
+			}
+		}
+	}
+
+	result.ServiceNameToServices = services
+	result.SecretNameToSNIs = secretToSNIs
+	return result
+}
+
+func knativeSelectSplit(splits []knative.IngressBackendSplit) knative.IngressBackendSplit {
+	if len(splits) == 0 {
+		return knative.IngressBackendSplit{}
+	}
+	res := splits[0]
+	maxPercentage := splits[0].Percent
+	if len(splits) == 1 {
+		return res
+	}
+	for i := 1; i < len(splits); i++ {
+		if splits[i].Percent > maxPercentage {
+			res = splits[i]
+			maxPercentage = res.Percent
+		}
+	}
+	return res
+}
